Refactor vision to not use an enum
