+++ src/main/java/frc/robot/Constants.java
@@ -72,23 +72,25 @@ public final class Constants {
 
   // CONTROL CONSTANT CLASSES
 
-  public static class Vision {
-    // public static final String kCameraName = "HD_Webcam_C615";
-    // Cam mounted facing forward, half a meter forward of center, half a meter up
-    // from center.
-    public static final Transform3d kRobotToCam = new Transform3d(new Translation3d(0.127, .3620516, .2397125),
-        new Rotation3d(0, 0, 0));
+  public static class BottomCamera {
+    public static String kName = "bottom";
+    public static Rotation3d kCameraRotation = new Rotation3d(Math.toRadians(0), Math.toRadians(0), Math.toRadians(30));
+    public static Translation3d kCameraOffset = new Translation3d(Units.inchesToMeters(9.25),
+        Units.inchesToMeters(-10.125), Units.inchesToMeters(18));
+    public static Matrix<N3, N1> kSingleTagDeviations = VecBuilder.fill(.9, .9, 6);
+    public static Matrix<N3, N1> kMultiTagDeviation = VecBuilder.fill(0.5, 0.5, 1);
 
-    // The layout of the AprilTags on the field
-    public static final AprilTagFieldLayout kTagLayout = AprilTagFieldLayout
-        .loadField(AprilTagFields.k2025ReefscapeWelded);
+  };
 
-    // The standard deviations of our vision estimated poses, which affect
-    // correction rate
-    // (Fake values. Experiment and determine estimation noise on an actual robot.)
-    public static final Matrix<N3, N1> kSingleTagStdDevs = VecBuilder.fill(4, 4, 8);
-    public static final Matrix<N3, N1> kMultiTagStdDevs = VecBuilder.fill(0.5, 0.5, 1);
-  }
+  public static class TopCamera {
+    public static String kName = "top";
+    public static Rotation3d kCameraRotation = new Rotation3d(0, Math.toRadians(0), Math.toRadians(180));
+    public static Translation3d kCameraOffset = new Translation3d(Units.inchesToMeters(-7), Units.inchesToMeters(0),
+        Units.inchesToMeters(14.75));
+    public static Matrix<N3, N1> kSingleTagDeviations = VecBuilder.fill(1.8, 1.8, 8);
+    public static Matrix<N3, N1> kMultiTagDeviation = VecBuilder.fill(0.5, 0.5, 1);
+
+  };
 
   public static final class AutonConstants {
